+++ src/main/java/cart/controller/ViewController.java
@@ -2,8 +2,8 @@ package cart.controller;
 
 import cart.domain.member.service.MemberService;
 import cart.domain.product.service.ProductService;
-import cart.dto.MemberResponse;
-import cart.dto.ProductResponse;
+import cart.domain.member.dto.MemberResponse;
+import cart.domain.product.dto.ProductResponse;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
