+++ src/main/java/cart/domain/cart/dto/CartCreateRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.domain.cart.dto;
 
 public class CartCreateRequest {
