+++ src/main/java/cart/domain/cart/service/CartService.java
@@ -6,9 +6,9 @@ import cart.dao.MemberDao;
 import cart.domain.member.entity.Member;
 import cart.dao.ProductDao;
 import cart.domain.product.entity.Product;
-import cart.dto.CartCreateRequest;
-import cart.dto.CartCreateResponse;
-import cart.dto.CartResponse;
+import cart.domain.cart.dto.CartCreateRequest;
+import cart.domain.cart.dto.CartCreateResponse;
+import cart.domain.cart.dto.CartResponse;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
