+++ src/main/java/cart/domain/member/dto/MemberInformation.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.domain.member.dto;
 
 public class MemberInformation {
