+++ src/main/java/cart/domain/member/dto/MemberResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.domain.member.dto;
 
 import cart.domain.member.entity.Member;
