+++ src/main/java/cart/domain/member/service/MemberService.java
@@ -2,9 +2,9 @@ package cart.domain.member.service;
 
 import cart.dao.MemberDao;
 import cart.domain.member.entity.Member;
-import cart.dto.MemberCreateRequest;
-import cart.dto.MemberCreateResponse;
-import cart.dto.MemberResponse;
+import cart.domain.member.dto.MemberCreateRequest;
+import cart.domain.member.dto.MemberCreateResponse;
+import cart.domain.member.dto.MemberResponse;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
