+++ src/main/java/cart/domain/product/dto/ProductResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.domain.product.dto;
 
 import cart.domain.product.entity.Product;
 import java.time.LocalDateTime;
