+++ src/test/java/cart/controller/CartControllerTest.java
@@ -13,8 +13,8 @@ import cart.common.auth.AuthInfoArgumentResolver;
 import cart.common.auth.AuthenticationCheckInterceptor;
 import cart.domain.auth.service.AuthService;
 import cart.domain.cart.service.CartService;
-import cart.dto.MemberInformation;
-import cart.dto.CartResponse;
+import cart.domain.member.dto.MemberInformation;
+import cart.domain.cart.dto.CartResponse;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
