+++ src/test/java/cart/domain/cart/service/CartServiceTest.java
@@ -13,8 +13,8 @@ import cart.dao.MemberDao;
 import cart.domain.member.entity.Member;
 import cart.dao.ProductDao;
 import cart.domain.product.entity.Product;
-import cart.dto.CartCreateRequest;
-import cart.dto.CartCreateResponse;
+import cart.domain.cart.dto.CartCreateRequest;
+import cart.domain.cart.dto.CartCreateResponse;
 import java.time.LocalDateTime;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
