+++ src/test/java/cart/dto/ProductCreateRequestTest.java
@@ -2,6 +2,7 @@ package cart.dto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import cart.domain.product.dto.ProductCreateRequest;
 import java.util.Optional;
 import javax.validation.ConstraintViolation;
 import javax.validation.Validation;
